+++ tests/integrationtests/test_oqt.py
@@ -1,19 +1,46 @@
 import asyncio
 import os
 import unittest
+from unittest import mock
 
 import geojson
+import pytest
 
 from ohsome_quality_analyst import oqt
 from ohsome_quality_analyst.api.request_models import (
     IndicatorDataRequest,
     IndicatorRequest,
-    ReportRequest,
 )
 from tests.conftest import FIXTURE_DIR
-from tests.integrationtests.utils import get_geojson_fixture
+from tests.integrationtests.utils import oqt_vcr
 
-from .utils import oqt_vcr
+
+@oqt_vcr.use_cassette
+@pytest.mark.parametrize(
+    "indicator,topic",
+    [
+        ("minimal", "topic_minimal"),
+        ("mapping-saturation", "topic_building_count"),
+        ("currentness", "topic_building_count"),
+        ("attribute-completeness", "topic_building_count"),
+    ],
+)
+def test_create_indicator(feature, indicator, topic, request):
+    """Minimal viable request for a single bpoly."""
+    topic = request.getfixturevalue(topic)
+    indicator = asyncio.run(oqt.create_indicator(indicator, feature, topic))
+    assert indicator.result.label is not None
+    assert indicator.result.value is not None
+    assert indicator.result.description is not None
+    assert indicator.result.figure is not None
+
+
+@oqt_vcr.use_cassette
+def test_create_report(feature):
+    report = asyncio.run(oqt.create_report("minimal", feature))
+    assert report.result.label is not None
+    assert report.result.class_ is not None
+    assert report.result.description is not None
 
 
 class TestOqt(unittest.TestCase):
@@ -30,85 +57,41 @@ class TestOqt(unittest.TestCase):
         with open(path, "r") as f:
             self.feature = geojson.load(f)
 
-    def run_tests(self, indicator):
-        self.assertIsNotNone(indicator.result.label)
-        self.assertIsNotNone(indicator.result.value)
-        self.assertIsNotNone(indicator.result.description)
-        self.assertIsNotNone(indicator.result.svg)
-
-    @oqt_vcr.use_cassette()
-    def test_create_indicator_bpolys(self):
-        """Test creating indicator from scratch."""
-        parameters = IndicatorRequest(topic=self.topic_key, bpolys=self.feature)
-        indicator = asyncio.run(oqt.create_indicator(parameters, self.indicator_name))
-        self.run_tests(indicator)
-
-    def test_create_indicator_not_implemented(self):
-        with self.assertRaises(NotImplementedError):
-            asyncio.run(oqt.create_indicator(""))
-
-    @oqt_vcr.use_cassette()
-    def test_create_report_bpolys(self):
-        """Test creating report from scratch using the 'bpolys'parameters ."""
-        parameters = ReportRequest(bpolys=self.feature)
-        report = asyncio.run(oqt.create_report(parameters, key=self.report_name))
-        self.assertIsNotNone(report.result.label)
-        self.assertIsNotNone(report.result.class_)
-        self.assertIsNotNone(report.result.description)
-
-    def test_create_report_not_implemented(self):
-        with self.assertRaises(NotImplementedError):
-            asyncio.run(oqt.create_report(""))
-
+    @mock.patch.dict("os.environ", {"OQT_GEOM_SIZE_LIMIT": "1"}, clear=True)
     def test_create_indicator_as_geojson_size_limit_bpolys(self):
-        feature = get_geojson_fixture("europe.geojson")  # TODO: use pytest fixture
         parameters = IndicatorRequest(
             topic=self.topic_key,
-            bpolys=feature,
+            bpolys=self.feature,
         )
         with self.assertRaises(ValueError):
             asyncio.run(
-                oqt.create_indicator_as_geojson(
-                    parameters, key=self.indicator_name, size_restriction=True
-                )
+                oqt.create_indicator_as_geojson(parameters, key=self.indicator_name)
             )
 
-    @oqt_vcr.use_cassette()
+    @mock.patch.dict("os.environ", {"OQT_GEOM_SIZE_LIMIT": "1"}, clear=True)
+    @oqt_vcr.use_cassette
     def test_create_indicator_as_geojson_size_limit_bpolys_ms(self):
         """Size limit is disabled for the Mapping Saturation indicator.
 
-        Test that no error gets raised
+        Test that no error gets raised.
         """
-        path = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "fixtures",
-            "heidelberg-altstadt-feature.geojson",
-        )
-        with open(path, "r") as f:
-            feature = geojson.load(f)
         parameters = IndicatorRequest(
             topic="building-count",
-            bpolys=feature,
+            bpolys=self.feature,
         )
         asyncio.run(
-            oqt.create_indicator_as_geojson(
-                parameters, key="mapping-saturation", size_restriction=True
-            )
+            oqt.create_indicator_as_geojson(parameters, key="mapping-saturation")
         )
 
-    @oqt_vcr.use_cassette()
+    @mock.patch.dict("os.environ", {"OQT_GEOM_SIZE_LIMIT": "1"}, clear=True)
+    @oqt_vcr.use_cassette
     def test_create_indicator_as_geojson_size_limit_topic_data(self):
-        """No error should be raised."""
-        path = os.path.join(
-            # TODO: use pytest fixture
-            os.path.dirname(os.path.abspath(__file__)),
-            "fixtures",
-            "europe.geojson",
-        )
-        with open(path, "r") as f:
-            feature = geojson.load(f)
+        """Size limit is disabled for request with custom data.
+
+        Test that no error gets raised.
+        """
         parameters = IndicatorDataRequest(
-            bpolys=feature,
+            bpolys=self.feature,
             topic={
                 "key": "key",
                 "name": "name",
@@ -124,9 +107,7 @@ class TestOqt(unittest.TestCase):
             },
         )
         asyncio.run(
-            oqt.create_indicator_as_geojson(
-                parameters, key="mapping-saturation", size_restriction=True
-            )
+            oqt.create_indicator_as_geojson(parameters, key="mapping-saturation")
         )